Single precision version of PyCLES!

+++ Csrc/entropies.h
@@ -2,21 +2,21 @@
 #include "parameters.h"
 #include "thermodynamic_functions.h"
 
-inline double sd_c(double pd, double T){
+inline float  sd_c(float  pd, float  T){
     return sd_tilde + cpd*log(T/T_tilde) -Rd*log(pd/p_tilde);
 }
 
-inline double sv_c(double pv, double T){
+inline float  sv_c(float  pv, float  T){
     return sv_tilde + cpv*log(T/T_tilde) - Rv * log(pv/p_tilde);
 }
 
-inline double sc_c(double L, double T){
+inline float  sc_c(float  L, float  T){
     return -L/T;
 }
 
-inline double s_tendency_c(double p0, double qt, double qv, double T, double qt_tendency, double T_tendency){
-    const double pv = pv_c(p0, qt, qv);
-    const double pd = pd_c(p0, qt, qv);
+inline float  s_tendency_c(float  p0, float  qt, float  qv, float  T, float  qt_tendency, float  T_tendency){
+    const float  pv = pv_c(p0, qt, qv);
+    const float  pd = pd_c(p0, qt, qv);
     return cpm_c(qt) * T_tendency / T +  (sv_c(pv,T) - sd_c(pd,T)) * qt_tendency;
 }
