+++ Csrc/lookup.h
@@ -3,21 +3,21 @@
 
 struct LookupStruct{
     size_t n_table;
-    double y_max;
-    double y_min;
-    double x_max;
-    double x_min;
-    double dx;
-    double dxi;
-    double* y;
+    float  y_max;
+    float  y_min;
+    float  x_max;
+    float  x_min;
+    float  dx;
+    float  dxi;
+    float * y;
 
 };
 
-void init_table(struct LookupStruct *LT, size_t n, double* x, double* y){
+void init_table(struct LookupStruct *LT, size_t n, float * x, float * y){
 
 
     /// Allocate memory required for storing the table
-    LT->y = calloc(n,sizeof(double));
+    LT->y = calloc(n,sizeof(float ));
 
     /// Determine bounds and set array table values
     LT->x_min = 9e16;
@@ -45,10 +45,10 @@ void free_table(struct LookupStruct *LT){
     free(LT->y);
 };
 
-inline double lookup(struct LookupStruct *LT, double x){
+inline float  lookup(struct LookupStruct *LT, float  x){
     const size_t indx = floor((x - LT->x_min)*LT->dxi);
-    const double y1 = LT->y[indx] ;
-    const double y2 = LT->y[indx + 1];
-    const double x1 = LT->x_min + LT-> dx * indx ;
+    const float  y1 = LT->y[indx] ;
+    const float  y2 = LT->y[indx + 1];
+    const float  x1 = LT->x_min + LT-> dx * indx ;
     return y1 + (x - x1) * (y2 - y1)*LT->dxi;
 };
